Changed to Greedy
Chose to do Greedy instead of Hill Climb. Did research on how to implement the search. Complete the method the cacluates the number of numbers that are in the wrong spot in the grid. Also, completed the documentation for the search.

Learn: manhattan won't be usful for my search.

RoadBlocks: making the neighbors and  setting there values.